+++ packages/integrations/actions/picoActions.js
@@ -1,6 +1,7 @@
 import { createAction } from '@irvingjs/core/actions';
 import {
   PICO_LOADED,
+  PICO_SCRIPT_ADDED,
   RECEIVE_PICO_PLAN_UPGRADE,
   RECEIVE_PICO_VERIFICATION_FAILURE,
   REQUIRE_UPGRADE_FOR_CORAL_SSO,
@@ -14,6 +15,14 @@ import {
  */
 export function actionPicoLoaded() {
   return createAction(PICO_LOADED);
+}
+
+/**
+ * Create an action that updates the store when the Pico script is added.
+ * @returns {{type, payload}} The Redux action.
+ */
+export function actionPicoScriptAdded() {
+  return createAction(PICO_SCRIPT_ADDED);
 }
 
 /**