+++ packages/integrations/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0-alpha.19](https://github.com/alleyinteractive/irving/packages/integrations/compare/v4.2.0-alpha.18...v4.2.0-alpha.19) (2020-09-30)
+
+**Note:** Version bump only for package @irvingjs/integrations
+
+
+
+
+
+# [4.2.0-alpha.17](https://github.com/alleyinteractive/irving/packages/integrations/compare/v4.2.0-alpha.16...v4.2.0-alpha.17) (2020-09-30)
+
+
+### Bug Fixes
+
+* **gtm:** fix XSS issue in GTM ([3731639](https://github.com/alleyinteractive/irving/packages/integrations/commit/3731639aefbe8e766b09b3d00fe030148e5dbc02))
+
+
+
+
+
 # [4.2.0-alpha.16](https://github.com/alleyinteractive/irving/packages/integrations/compare/v4.2.0-alpha.15...v4.2.0-alpha.16) (2020-09-28)